feat(llms-usage): route usage tracking through JOBS_MASTER-style messaging into tyr
LLMSUsageCrudServiceImpl no longer writes to Mongo directly on the calling
chat-completion thread. It now emits a LLMUsageDetailPayload through the
internal message broker (LLMS-USAGE-MONITOR/USAGE-CONCENTRATOR), matching the
pattern already used for workflow/job-status consolidation.

The receiving side - LLMUsageDetail/LLMDailyUsageDetail, their repositories,
the daily consolidation job, the drill-down aggregation service, and the two
admin/user reporting controllers - moves into a new
gebo.architecture.compute.workflow package (ai.gebo.architecture.llms.usage),
off gebo.ragsystem.client.rest and gebo.architecture.llms.abstraction.layer.
This shifts the usage-reporting REST surface off brain onto tyr (+ the
monolith), mirroring the earlier job-status/workflow-stats move.

ILLMSUsageCrudService.enqueueUsage now takes a DTO instead of the entity, and
ModelType moves to ai.gebo.model (gebo.base.model) so it stays shared across
the emitter, payload, and receiver-side entity without a circular dependency.
Also fixes a pre-existing bug where the computed chat latency was never
actually set on the outgoing usage record.
