+++ gebo.architecture.parent/gebo.architecture.compute.workflow/pom.xml
@@ -42,5 +42,10 @@
 			<groupId>org.springframework.security</groupId>
 			<artifactId>spring-security-core</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>ai.gebo.architecture</groupId>
+			<artifactId>gebo.architecture.security</artifactId>
+			<version>${project.version}</version>
+		</dependency>
 	</dependencies>
 </project>
