+++ gebo.architecture.parent/gebo.architecture.compute.workflow/src/main/java/ai/gebo/architecture/llms/usage/model/LLMUsageAggregationBucket.java
@@ -1,6 +1,6 @@
-package ai.gebo.llms.chat.client.rest.model;
+package ai.gebo.architecture.llms.usage.model;
 
-import ai.gebo.llms.abstraction.layer.model.ModelType;
+import ai.gebo.model.ModelType;
 import lombok.Data;
 
 /**
