+++ gebo.architecture.parent/gebo.architecture.compute.workflow/src/main/java/ai/gebo/architecture/llms/usage/model/LLMUsageDetail.java
@@ -1,32 +1,29 @@
-package ai.gebo.llms.abstraction.layer.model;
+package ai.gebo.architecture.llms.usage.model;
 
 import java.util.UUID;
 
 import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.index.HashIndexed;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import ai.gebo.model.ModelType;
 import lombok.Data;
 
 @Data
 @Document
-public class LLMDailyUsageDetail {
+public class LLMUsageDetail {
 	@Id
 	private String id = UUID.randomUUID().toString();
+
 	private String providerId;
 	private String username;
 	private String model;
 	private String callerStack;
 	private ModelType modelType;
-	private int year;
-	private int month;
-	private int day;
-	private long latencyMin;
-	private long latencyMax;
-	private long latencyAvg;
+	private long latency;
 	private long inputToken;
 	private long outputToken;
 	private long totalToken;
-	private long nrRequests;
-	
-
+	@HashIndexed
+	private long timestamp = System.currentTimeMillis();
 }
