+++ gebo.architecture.parent/gebo.architecture.compute.workflow/src/main/java/ai/gebo/architecture/llms/usage/model/LLMUsageDrillDownResult.java
@@ -1,4 +1,4 @@
-package ai.gebo.llms.chat.client.rest.model;
+package ai.gebo.architecture.llms.usage.model;
 
 import java.util.List;
