--- gebo.architecture.parent/gebo.architecture.llms.abstraction.layer/src/main/java/ai/gebo/llms/abstraction/layer/model/ModelType.java
@@ -1,5 +1,0 @@
-package ai.gebo.llms.abstraction.layer.model;
-
-public enum ModelType {
-	CHAT, EMBEDDING, IMAGE, RANKER, TTS, TRANSCRIPT
-}
