+++ gebo.architecture.parent/gebo.architecture.llms.abstraction.layer/src/main/java/ai/gebo/llms/abstraction/layer/services/ILLMSUsageCrudService.java
@@ -1,10 +1,7 @@
 package ai.gebo.llms.abstraction.layer.services;
 
-import org.springframework.scheduling.annotation.Async;
-
-import ai.gebo.llms.abstraction.layer.model.LLMUsageDetail;
+import ai.gebo.llms.abstraction.layer.dto.LLMUsageDetailDto;
 
 public interface ILLMSUsageCrudService {
-	@Async
-	public void enqueueUsage(LLMUsageDetail usage);
+	public void enqueueUsage(LLMUsageDetailDto usage);
 }
