+++ gebo.architecture.parent/gebo.architecture.llms.abstraction.layer/src/main/java/ai/gebo/llms/abstraction/layer/services/impl/LLMSUsageCrudServiceImpl.java
@@ -1,137 +1,70 @@
 package ai.gebo.llms.abstraction.layer.services.impl;
 
-import java.time.Instant;
-import java.time.LocalDate;
-import java.time.ZoneId;
-import java.util.GregorianCalendar;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.stream.Stream;
+import java.util.List;
 
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
-import ai.gebo.llms.abstraction.layer.model.LLMDailyUsageDetail;
-import ai.gebo.llms.abstraction.layer.model.LLMUsageDetail;
-import ai.gebo.llms.abstraction.layer.model.ModelType;
-import ai.gebo.llms.abstraction.layer.repository.LLMDailyUsageDetailRepository;
-import ai.gebo.llms.abstraction.layer.repository.LLMUsageDetailRepository;
+import ai.gebo.application.messaging.IGMessageBroker;
+import ai.gebo.application.messaging.IGMessageEmitter;
+import ai.gebo.application.messaging.IMessageEnvelopeFactory;
+import ai.gebo.application.messaging.SystemComponentType;
+import ai.gebo.application.messaging.model.GMessageEnvelope;
+import ai.gebo.application.messaging.model.GStandardModulesConstraints;
+import ai.gebo.core.messages.LLMUsageDetailPayload;
+import ai.gebo.llms.abstraction.layer.dto.LLMUsageDetailDto;
 import ai.gebo.llms.abstraction.layer.services.ILLMSUsageCrudService;
 import lombok.AllArgsConstructor;
 
+/**
+ * Emits usage events to the {@code LLMS-USAGE-MONITOR}/{@code USAGE-CONCENTRATOR}
+ * receiver hosted in {@code gebo.architecture.compute.workflow} (package
+ * {@code ai.gebo.architecture.llms.usage}), rather than persisting directly - the
+ * receiver's own dedicated thread does the Mongo write and the periodic
+ * consolidation, keeping both off the calling chat-completion thread.
+ */
 @Service
 @AllArgsConstructor
-public class LLMSUsageCrudServiceImpl implements ILLMSUsageCrudService {
-	private final LLMUsageDetailRepository usageRepo;
-	private final LLMDailyUsageDetailRepository consolidatedRepo;
+public class LLMSUsageCrudServiceImpl implements ILLMSUsageCrudService, IGMessageEmitter {
+	private final IGMessageBroker broker;
+	private final IMessageEnvelopeFactory envelopeFactory;
 
 	@Override
-	@Async
-	public void enqueueUsage(LLMUsageDetail usage) {
-		usageRepo.insert(usage);
+	public void enqueueUsage(LLMUsageDetailDto usage) {
+		LLMUsageDetailPayload payload = new LLMUsageDetailPayload();
+		payload.setProviderId(usage.getProviderId());
+		payload.setUsername(usage.getUsername());
+		payload.setModel(usage.getModel());
+		payload.setCallerStack(usage.getCallerStack());
+		payload.setModelType(usage.getModelType());
+		payload.setLatency(usage.getLatency());
+		payload.setInputToken(usage.getInputToken());
+		payload.setOutputToken(usage.getOutputToken());
+		payload.setTotalToken(usage.getTotalToken());
+		payload.setUsageTimestamp(System.currentTimeMillis());
+
+		GMessageEnvelope<LLMUsageDetailPayload> envelope = envelopeFactory.newMessageFrom(this, payload);
+		envelope.setTargetModule(GStandardModulesConstraints.LLMS_USAGE_MONITOR);
+		envelope.setTargetComponent(GStandardModulesConstraints.USAGE_CONCENTRATOR);
+		broker.accept(envelope);
 	}
 
-	@Scheduled(initialDelay = 5000, fixedRate = 10 * 60 * 1000)
-	public void consolidateTick() {
-		ZoneId zone = ZoneId.systemDefault();
-		LocalDate today = LocalDate.now(zone);
-
-		// Only completed days are consolidated: everything up to the last
-		// millisecond of yesterday.
-		long lastMillisOfYesterday = today.atStartOfDay(zone).toInstant().toEpochMilli() - 1;
-		long todayFirstMillisecond = lastMillisOfYesterday + 1;
-		long todayLastMillisecond = today.plusDays(1).atStartOfDay(zone).toInstant().toEpochMilli() - 1;
-		;
-		// Aggregate the raw values grouping by
-		// providerId, username, model, callerStack, modelType, year, month, day.
-		Map<ConsolidationKey, DailyAccumulator> grouped = new HashMap<>();
-		try (Stream<LLMUsageDetail> stream = usageRepo.findByTimestampGreaterThanEqualAndTimestampLessThanEqual(
-				todayFirstMillisecond, todayLastMillisecond)) {
-			stream.forEach(detail -> {
-				LocalDate date = Instant.ofEpochMilli(detail.getTimestamp()).atZone(zone).toLocalDate();
-				ConsolidationKey key = new ConsolidationKey(detail.getProviderId(), detail.getUsername(),
-						detail.getModel(), detail.getCallerStack(), detail.getModelType(), date.getYear(),
-						date.getMonthValue(), date.getDayOfMonth());
-				grouped.computeIfAbsent(key, k -> new DailyAccumulator()).add(detail);
-			});
-		}
-		if (grouped.isEmpty()) {
-			return;
-		}
-
-		// Complete the sums of the target daily documents, merging into any
-		// previously consolidated record for the same key.
-		for (Map.Entry<ConsolidationKey, DailyAccumulator> entry : grouped.entrySet()) {
-			ConsolidationKey key = entry.getKey();
-			LLMDailyUsageDetail target = consolidatedRepo
-					.findByProviderIdAndUsernameAndModelAndCallerStackAndModelTypeAndYearAndMonthAndDay(
-							key.providerId(), key.username(), key.model(), key.callerStack(), key.modelType(),
-							key.year(), key.month(), key.day())
-					.orElseGet(() -> newDailyUsageDetail(key));
-			entry.getValue().mergeInto(target);
-			consolidatedRepo.save(target);
-		}
-
-		// The processed raw details have been folded into the daily documents,
-		// remove every detail belonging to an already-consolidated day.
-		usageRepo.deleteByTimestampLessThanEqual(lastMillisOfYesterday);
+	@Override
+	public String getMessagingModuleId() {
+		return GStandardModulesConstraints.LLMS_USAGE_MONITOR;
 	}
 
-	private static LLMDailyUsageDetail newDailyUsageDetail(ConsolidationKey key) {
-		LLMDailyUsageDetail daily = new LLMDailyUsageDetail();
-		daily.setProviderId(key.providerId());
-		daily.setUsername(key.username());
-		daily.setModel(key.model());
-		daily.setCallerStack(key.callerStack());
-		daily.setModelType(key.modelType());
-		daily.setYear(key.year());
-		daily.setMonth(key.month());
-		daily.setDay(key.day());
-		return daily;
+	@Override
+	public String getMessagingSystemId() {
+		return GStandardModulesConstraints.USAGE_CONCENTRATOR;
 	}
 
-	private record ConsolidationKey(String providerId, String username, String model, String callerStack,
-			ModelType modelType, int year, int month, int day) {
+	@Override
+	public SystemComponentType getComponentType() {
+		return SystemComponentType.APPLICATION_COMPONENT;
 	}
 
-	private static final class DailyAccumulator {
-		private long inputToken;
-		private long outputToken;
-		private long totalToken;
-		private long nrRequests;
-		private long latencySum;
-		private long latencyMin = Long.MAX_VALUE;
-		private long latencyMax = Long.MIN_VALUE;
-
-		void add(LLMUsageDetail detail) {
-			inputToken += detail.getInputToken();
-			outputToken += detail.getOutputToken();
-			totalToken += detail.getTotalToken();
-			nrRequests++;
-			latencySum += detail.getLatency();
-			latencyMin = Math.min(latencyMin, detail.getLatency());
-			latencyMax = Math.max(latencyMax, detail.getLatency());
-		}
-
-		void mergeInto(LLMDailyUsageDetail target) {
-			long existingNr = target.getNrRequests();
-			long existingLatencySum = target.getLatencyAvg() * existingNr;
-			long combinedNr = existingNr + nrRequests;
-
-			target.setInputToken(target.getInputToken() + inputToken);
-			target.setOutputToken(target.getOutputToken() + outputToken);
-			target.setTotalToken(target.getTotalToken() + totalToken);
-			target.setNrRequests(combinedNr);
-			target.setLatencyAvg(combinedNr > 0 ? (existingLatencySum + latencySum) / combinedNr : 0);
-			if (existingNr > 0) {
-				target.setLatencyMin(Math.min(target.getLatencyMin(), latencyMin));
-				target.setLatencyMax(Math.max(target.getLatencyMax(), latencyMax));
-			} else {
-				target.setLatencyMin(latencyMin);
-				target.setLatencyMax(latencyMax);
-			}
-		}
+	@Override
+	public List<String> getEmittedPayloadTypes() {
+		return List.of(LLMUsageDetailPayload.class.getName());
 	}
-
 }
