+++ gebo.core.parent/gebo.base.model/src/main/java/ai/gebo/model/ModelType.java
@@ -1,0 +1,14 @@
+/**
+ * This Source Code is subject to the terms of the
+ * Gebo.ai community version Mozilla Public License Version 2.0 (MPL-2.0) — With Data Protection Clauses
+ * If a copy of the LICENCE was not distributed with this file, You can obtain one at
+ * https://gebo.ai/gebo-ai-community-version-mozilla-public-license-version-2-0-mpl-2-0-with-data-protection-clauses/
+ * and https://mozilla.org/MPL/2.0/.
+ * Copyright (c) 2025+ Gebo.ai
+ */
+
+package ai.gebo.model;
+
+public enum ModelType {
+	CHAT, EMBEDDING, IMAGE, RANKER, TTS, TRANSCRIPT
+}
